Update Get Posts Success
- Updated the get posts success function to now check if the user's email matches the post owner's email

There was a bug where two people with the same name but different accounts would see the edit/delete buttons on posts that weren't theirs. To fix this, I changed the returned data to return the whole object and not just their name so the client can check the emails as the emails are unique.